Factor out interaction indices, register GemNet-OC
