+++ ocpmodels/models/__init__.py
@@ -9,6 +9,7 @@ from .dimenet import DimeNetWrap as DimeNet
 from .dimenet_plus_plus import DimeNetPlusPlusWrap as DimeNetPlusPlus
 from .forcenet import ForceNet
 from .gemnet.gemnet import GemNetT
+from .gemnet_oc.gemnet_oc import GemNetOC
 from .painn.painn import PaiNN
 from .schnet import SchNetWrap as SchNet
 from .spinconv import spinconv
